+++ @ornikar/storybook-config/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.3](https://github.com/ornikar/shared-configs/compare/@ornikar/storybook-config@3.3.2...@ornikar/storybook-config@3.3.3) (2022-01-14)
+
+**Note:** Version bump only for package @ornikar/storybook-config
+
+
+
+
+
 ## [3.3.2](https://github.com/ornikar/shared-configs/compare/@ornikar/storybook-config@3.3.1...@ornikar/storybook-config@3.3.2) (2021-12-09)
 
 **Note:** Version bump only for package @ornikar/storybook-config